controllers refactor
